Third commit
Fixed that player cant double down after he drew the first card.
Added that you can bet on games and get returns on them.

Need to test if split mode is returning the right amount of money.
Money should not go negative.
Need to add some visuals to cards that it looks more like a game of blackjack.