Sixth commit
Money cant go negative, player has to either match the bet that he inputed or leave the game. If he wont match the bet that he inputed then the program keeps asking to do so.

Started to change up the print functions a bit, so that these functions print out cards Not dictioneries.